ppt-parse/requirements-locked.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.14
botocore==1.42.14
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.125.0
h11==0.16.0
httptools==0.7.1
idna==3.11
jmespath==1.0.1
lxml==6.0.2
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-magic==0.4.27
python-multipart==0.0.20
python-pptx==1.0.2
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xlsxwriter==3.2.9
